refactor(utils): Утилиты для работы с константными перечислениями перенесены в enums
- **valuesAsKeys** и подобные функции вынесены в enums/utils
- Добавлено константное перечисление **ORIENTATIONS**
