+++ packages/core/enums/index.ts
@@ -1,4 +1,6 @@
+export * from './utils';
 export * from './keys';
 export * from './QxDevice';
 export * from './QxSize';
 export * from './QxBorderSize';
+export * from './orientations';
