+++ packages/core/enums/orientations.ts
@@ -1,0 +1,9 @@
+import { Values } from '@core/types';
+import { valuesAsKeysFromArray } from './utils';
+
+export const ORIENTATIONS = valuesAsKeysFromArray([
+    'horizontal',
+    'vertical',
+]);
+
+export type Orientation = Values<typeof ORIENTATIONS>;
