+++ packages/core/index.ts
@@ -2,7 +2,7 @@ import './polyfills';
 
 // WARNING: Изменение порядка экспорта может сломать сборку
 export * from './types';
-export * from './utils';
 export * from './enums';
+export * from './utils';
 export * from './styles';
 export * from './src';
