+++ packages/core/utils/hooks/useKeyboardNavigation/constants.ts
@@ -1,13 +1,8 @@
-import { valuesAsKeysFromArray } from '../../object';
+import { valuesAsKeysFromArray } from '@core/enums';
 
 export const ARROW_KEYS = valuesAsKeysFromArray([
     'ArrowUp',
     'ArrowDown',
     'ArrowLeft',
     'ArrowRight',
-]);
-
-export const ORIENTATIONS = valuesAsKeysFromArray([
-    'horizontal',
-    'vertical',
 ]);
