+++ packages/core/utils/object/index.ts
@@ -1,4 +1,3 @@
-export * from './valuesAsKeys';
 export { getKeys } from './getKeys';
 export * from './omitProps';
 export * from './mapObjectKeys';
