+++ packages/core/utils/pointers.ts
@@ -1,4 +1,4 @@
-import { valuesAsKeysFromArray } from './object/valuesAsKeys';
+import { valuesAsKeysFromArray } from '@core/enums';
 
 export const POINTERS_TYPE = valuesAsKeysFromArray([
     'mouse',
